+++ zio-query/shared/src/main/scala/zio/query/Described.scala
