+++ zio-query/shared/src/main/scala/zio/query/QueryFailure.scala
