+++ zio-query/shared/src/main/scala/zio/query/ZQuery.scala
@@ -3,6 +3,8 @@ package zio.query
 import zio._
 import zio.clock._
 import zio.duration._
+import zio.query.internal.{ BlockedRequest, BlockedRequests, Continue, Result }
+
 
 /**
  * A `ZQuery[R, E, A]` is a purely functional description of an effectual query
